vertex_ai/transak_i1/pyproject.toml
[project]
name = "transak-i1-keras-functional-api"
version = "0.1.0"
description = "Classification of Artur's Revolut transactions with Deep Neural Network using the Keras Functional API with TensorFlow."
# We specify Python 3.10.x is required.
requires-python = ">=3.10, <3.11"

dependencies = [
    "tensorflow==2.17.0",
    "numpy==1.26.4",
    "pandas==2.2.2",
    "matplotlib==3.9.0",
    "scikit-learn==1.5.0",
    "seaborn==0.13.2",
    "jupyterlab==4.2.1",
    "ipykernel==6.29.4",
    "google-cloud-bigquery[pandas]==3.24.0",
    "google-cloud-bigquery-storage==2.31.0",
    "google-cloud-aiplatform[autologging]==1.100.0",
    "pydot==4.0.1",
    "graphviz==0.21",
]

[project.optional-dependencies]
dev = [
    "black==24.4.2",
    "ruff==0.4.8",
    "mypy==1.10.0",
]
pipelines = [
    "kfp==2.7.0",
    "google-cloud-pipeline-components==2.20.1"
]
[tool.setuptools.packages.find]
# [tool.setuptools]
# packages = ["common", "data_prep", "evaluation", "pipeline", "trainer"]
[tool.uv]
# This section is optional but can be used for uv-specific configurations
